backend/requirements.txt
# ─────────────────────────────────────────────────────
# Core Framework
# ─────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
python-multipart>=0.0.9
pydantic[email]>=2.7.4
pydantic-settings>=2.3.4

# ─────────────────────────────────────────────────────
# Database & ORM
# ─────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.31
asyncpg>=0.29.0
aiosqlite>=0.20.0
alembic>=1.13.1
greenlet>=3.0.3

# ─────────────────────────────────────────────────────
# Authentication & Security
# ─────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.3
cryptography>=42.0.8
pyotp>=2.9.0

# ─────────────────────────────────────────────────────
# AI / NLP / Embeddings
# ─────────────────────────────────────────────────────
openai>=1.35.7
httpx>=0.27.0
tiktoken>=0.7.0

# ─────────────────────────────────────────────────────
# OCR
# ─────────────────────────────────────────────────────
pytesseract>=0.3.10
Pillow>=10.3.0
pdf2image>=1.17.0

# ─────────────────────────────────────────────────────
# Email Parsing
# ─────────────────────────────────────────────────────
mail-parser>=3.15.0

# ─────────────────────────────────────────────────────
# Vector Database
# ─────────────────────────────────────────────────────
chromadb>=0.5.0

# ─────────────────────────────────────────────────────
# Task Queue
# ─────────────────────────────────────────────────────
celery[redis]>=5.4.0
redis>=5.0.7

# ─────────────────────────────────────────────────────
# Telephony
# ─────────────────────────────────────────────────────
twilio>=9.2.1

# ─────────────────────────────────────────────────────
# Utilities
# ─────────────────────────────────────────────────────
python-dotenv>=1.0.1
structlog>=24.2.0
aiofiles>=24.1.0
tenacity>=8.4.2

# ─────────────────────────────────────────────────────
# Testing
# ─────────────────────────────────────────────────────
pytest>=8.2.2
pytest-asyncio>=0.23.7
pytest-cov>=5.0.0
